requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.3.5
    # via aiohttp
aiohttp==3.10.3
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via aiohttp
build==1.2.1
    # via pip-tools
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
filelock==3.15.4
    # via huggingface-hub
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.6.1
    # via huggingface-hub
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   langchain-mistralai
    #   openai
httpx-sse==0.4.0
    # via langchain-mistralai
huggingface-hub==0.24.5
    # via tokenizers
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.5.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.2.12
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.2.11
    # via -r requirements.in
langchain-core==0.2.29
    # via
    #   langchain
    #   langchain-community
    #   langchain-mistralai
    #   langchain-openai
    #   langchain-text-splitters
langchain-mistralai==0.1.12
    # via -r requirements.in
langchain-openai==0.1.21
    # via -r requirements.in
langchain-text-splitters==0.2.2
    # via langchain
langsmith==0.1.98
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.21.3
    # via dataclasses-json
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
openai==1.40.3
    # via
    #   -r requirements.in
    #   langchain-openai
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   build
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
pip-tools==7.4.1
    # via -r requirements.in
pydantic==2.8.2
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.20.1
    # via pydantic
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langsmith
    #   tiktoken
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.32
    # via
    #   langchain
    #   langchain-community
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.7.0
    # via langchain-openai
tokenizers==0.20.0
    # via langchain-mistralai
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   openai
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.2
    # via requests
wheel==0.44.0
    # via pip-tools
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
